third_party/WebKit/LayoutTests/editing/inserting/insert-into-table.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => {
  assert_selection(
    [
      '<div contenteditable><table><tbody><tr>',
        '<td>foo|\n   </td>',
      '</tr></tbody></table></div>'
    ].join(''),
    'insertHTML bar',
    [
      '<div contenteditable><table><tbody><tr>',
        '<td>foobar|\n   </td>',
      '</tr></tbody></table></div>'
    ].join(''));
}, 'insertHTML text into a table cell following white spaces should not crash');
</script>
